1 Фикс открытия приложения из уведомления: добавлен data scheme="player" в Манифест 2 Убрана  зависимость от MutableList вплоть до Playlist, где ему самое место 3 Убран атрибут requestLegacyExternalStorage из Манифеста. Теперь сохранение файла производится в Recordings (находится внутри каталога приложения внешнего хранилища) 4 Фикс: теперь запись работает при перевороте экрана 5 bound-сервис: запись голосовых сообщений работает в фоне и перемещена в отдельный пакет (в будущем возможен перенос в отдельный модуль) 6 Остановить запись теперь можно из уведомления (за счёт остановки сервиса и Pending Intent'а для запуска активити) 7 При переоткрытии приложения записывающий сервис "подхватывается" за счёт isMyServiceConnected(rebind) 8 Snackbars вместо toast 9 SingleLiveEvent.kt 10 Refactor: перенос создания модели заметки во ViewModel и создания путей в FileUtils.kt 11 Styling: тёмная тема

+++ app/src/main/java/com/alox1d/vkvoicenotes/data/model/VoiceNoteDTO.kt
@@ -10,7 +10,7 @@ import kotlinx.android.parcel.Parcelize
 @Entity
 @Parcelize
 data class VoiceNoteDTO(
-    @PrimaryKey(autoGenerate = true) var id: Int = 0,
+    @PrimaryKey(autoGenerate = true) var id: Long = 0,
     val name: String?,
     val path: String,
     val duration: String?,
